+++ src/main/typescript/item/wvItem.ts
@@ -76,18 +76,12 @@ export default class WvItem extends Item {
 
   /** Check whether the item has a compendium link in its flags. */
   get hasCompendiumLink(): boolean {
-    const sourceId = this.getFlag("core", "sourceId");
-    if (typeof sourceId !== "string") return false;
-
-    return SYSTEM_COMPENDIUM_SOURCE_ID_REGEX.test(sourceId);
+    return hasCompendiumLink(this);
   }
 
   /** Check whether the item has a compendium link that is currently enabled. */
   get hasEnabledCompendiumLink(): boolean {
-    return (
-      this.hasCompendiumLink &&
-      !this.getFlag(CONSTANTS.systemId, "disableCompendiumLink")
-    );
+    return hasEnabledCompendiumLink(this);
   }
 
   /** Check whether the item has the type of one of the prototype items. */
@@ -218,6 +212,25 @@ export type ItemFlags = {
   overwriteNotesWithCompendium?: boolean;
   overwriteRulesWithCompendium?: boolean;
 };
+
+/** Check whether the passed item has a compendium link in its flags. */
+export function hasCompendiumLink(item: foundry.documents.BaseItem) {
+  const sourceId = item.getFlag("core", "sourceId");
+  if (typeof sourceId !== "string") return false;
+
+  return SYSTEM_COMPENDIUM_SOURCE_ID_REGEX.test(sourceId);
+}
+
+/**
+ * Check whether the passed item has a compendium link that is currently
+ * enabled.
+ */
+export function hasEnabledCompendiumLink(item: foundry.documents.BaseItem) {
+  return (
+    hasCompendiumLink(item) &&
+    !item.getFlag(CONSTANTS.systemId, "disableCompendiumLink")
+  );
+}
 
 /** Fetch update data for an item from its compendium prototype. */
 export async function getUpdateDataFromCompendium(